pyproject.toml
[project]
name = "test-discord"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.5.2",
    "dotenv>=0.9.9",
    "openai-agents>=0.0.16",
]
